requirements.txt
gensim
pinecone-client
nltk
PyMuPDF